+++ versions/us1/mischiefmakers.yaml
@@ -114,6 +114,7 @@ segments:
       - [0x059EA0, c]
       - [0x05D120, c]
       - [0x05E230, c]
+      - [0x05EA30, c]
       - [0x060CF0, c]
       - [0x061B80, asm]
       - [0x066250, c]
